Refactor unit tests and add new attribute
Moved unit test methods from xUnit project to NUnit project. Deleted ParagraphPortionTests.cs in xUnit and transferred its content method to respective classes in the NUnit project. Added a new custom attribute - SlideParagraphAttribute in SlideParagraphAttribute.cs. It simplifies the passing of parameters for the test method - IndentLevel_Getter_returns_indent_level.
