+++ test/ShapeCrawler.Tests.Unit/Helpers/SlideParagraphAttribute.cs
@@ -1,0 +1,48 @@
+﻿using NUnit.Framework.Interfaces;
+using NUnit.Framework.Internal;
+using NUnit.Framework.Internal.Builders;
+
+namespace ShapeCrawler.Tests.Unit.Helpers;
+
+[AttributeUsage(AttributeTargets.Method, AllowMultiple = true)]
+public class SlideParagraphAttribute : Attribute, ITestBuilder
+{
+    private readonly string testName;
+    private readonly string pptxName;
+    private readonly int slideNumber;
+    private readonly string shapeName;
+    private readonly int paragraphNumber;
+    private readonly object expectedResult;
+
+    public SlideParagraphAttribute(
+        string testName,
+        string pptxName,
+        int slideNumber,
+        string shapeName,
+        int paragraphNumber,
+        object expectedResult)
+    {
+        this.testName = testName;
+        this.pptxName = pptxName;
+        this.slideNumber = slideNumber;
+        this.shapeName = shapeName;
+        this.paragraphNumber = paragraphNumber;
+        this.expectedResult = expectedResult;
+    }
+    
+    public IEnumerable<TestMethod> BuildFrom(IMethodInfo method, Test suite)
+    {
+        var pres = new Presentation(SCTest.StreamOf(this.pptxName));
+        var paragraph = pres.Slides[this.slideNumber - 1].Shapes.GetByName(this.shapeName).TextFrame
+            .Paragraphs[this.paragraphNumber - 1];
+
+        var parameters = new TestCaseParameters(new[] { paragraph, this.expectedResult });
+
+        if (!string.IsNullOrEmpty(this.testName))
+        {
+            parameters.TestName = this.testName;
+        }
+
+        yield return new NUnitTestCaseBuilder().BuildTestMethod(method, suite, parameters);
+    }
+}
